backend/kpi_app/serializers.py
from rest_framework import serializers

from .models import Circle, KPI


class CircleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Circle
        fields = ['id', 'name']


class KPISerializer(serializers.ModelSerializer):
    class Meta:
        model = KPI
        fields = ['id', 'name', 'range_id', 'periodicity_id']