backend/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.5.0
